Fix signed operations on file reading

+++ filereading.hpp
@@ -2,7 +2,6 @@
 #define FILEREADING_H
 
 #include <fstream>
-#include <iostream>
 #include <istream>
 
 using ENDIAN = std::endian;
@@ -31,10 +30,10 @@ using ENDIAN = std::endian;
 
 // Read 32 bit signed integer from file
 [[nodiscard]] inline long int readInt32(std::ifstream &src, ENDIAN endianness) {
-  int data;
+  unsigned int data;
   src.read(reinterpret_cast<char *>(&data), 4);
   if (ENDIAN::native != endianness) {
-    return _byteswap_ulong(data);
+    return static_cast<int>(_byteswap_ulong(data));
   }
   return data;
 }
@@ -42,10 +41,10 @@ using ENDIAN = std::endian;
 // Read 16 bit signed integer from file
 [[nodiscard]] inline short int readInt16(std::ifstream &src,
                                          ENDIAN endianness) {
-  short int data;
+  unsigned short int data;
   src.read(reinterpret_cast<char *>(&data), 2);
   if (ENDIAN::native != endianness) {
-    return _byteswap_ushort(data);
+    return static_cast<short int>(_byteswap_ushort(data));
   }
   return data;
 }
